samples: mcumgr: ble_mcumgr: Move BLE MCUmgr code to separate library
Allows for re-use in applications
