+++ samples/mcumgr/ble_mcumgr/prj.conf
@@ -39,6 +39,7 @@ CONFIG_MCUMGR_GRP_OS_MCUMGR_PARAMS=y
 CONFIG_MCUMGR_GRP_OS_BOOTLOADER_INFO=y
 CONFIG_NRF_SDH_BLE=y
 CONFIG_NRF_SDH=y
+CONFIG_BLE_MCUMGR=y
 
 CONFIG_LOG=y
 CONFIG_LOG_MODE_MINIMAL=y
