+++ subsys/bluetooth/services/CMakeLists.txt
@@ -9,3 +9,4 @@ add_subdirectory_ifdef(CONFIG_BLE_DIS ble_dis)
 add_subdirectory_ifdef(CONFIG_BLE_HRS ble_hrs)
 add_subdirectory_ifdef(CONFIG_BLE_LBS ble_lbs)
 add_subdirectory_ifdef(CONFIG_BLE_NUS ble_nus)
+add_subdirectory_ifdef(CONFIG_BLE_MCUMGR ble_mcumgr)
