+++ subsys/bluetooth/services/Kconfig
@@ -9,3 +9,4 @@ rsource "ble_dis/Kconfig"
 rsource "ble_hrs/Kconfig"
 rsource "ble_lbs/Kconfig"
 rsource "ble_nus/Kconfig"
+rsource "ble_mcumgr/Kconfig"
